Add ViewHelper middleware to package provider.
