+++ providers/AdonisAuthScaffoldProvider.js
@@ -28,6 +28,22 @@ class AdonisAuthScaffoldProvider extends ServiceProvider {
   }
 
   /**
+   * Setup scaffold command for authentications.
+   *
+   * @method _setupScaffoldCommand
+   *
+   * @return {void}
+   *
+   * @private
+   */
+  _setupViewHelperMiddleware() {
+    this.app.bind("AdonisAuthScaffold/Middleware/ViewHelper", (app) => {
+      const ViewHelper = require('../src/Middleware/ViewHelper')
+      return new ViewHelper()
+    });
+  }
+
+  /**
    * Registers IOC bindings for this package.
    *
    * @method register
@@ -36,6 +52,7 @@ class AdonisAuthScaffoldProvider extends ServiceProvider {
    */
   register() {
     this._setupScaffoldCommand();
+    this._setupViewHelperMiddleware();
   }
 
   /**
